proyecto/requirements.txt
# Geoespacial
geopandas>=0.14.0
shapely>=2.0.2
pyproj>=3.6.1
rasterio>=1.3.9
fiona>=1.9.5
osmnx>=1.7.1

# Data Science
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
scipy>=1.11.0

# Análisis Espacial
libpysal>=4.9.0
esda>=2.5.0
splot>=1.1.5.post1

# Visualización
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
folium>=0.15.0

# Web
streamlit>=1.28.0
streamlit-folium>=0.15.0

# Base de datos
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.20
geoalchemy2>=0.14.0

# Utilidades
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
